Finished assignment
Finished code for fibonacci, mergeSort and an additional assignment.
I had a lot of trouble with mergeSort, as I was able to merge arrays,
but I wasn't assigning variables correctly so the merged arrays weren't
returning correctly. The difference between the code I was stuck on and
the final product was a difference of 4 words. So frustrating! But a fun
challenge nonetheless. Onward!